requirements.txt
fastapi 

uvicorn 

openai 

python-multipart 

psycopg2 

pydantic



python-dotenv
google-generativeai